git subrepo pull (merge) OAuth2OOo
subrepo:
  subdir:   "OAuth2OOo"
  merged:   "d9115d2"
upstream:
  origin:   "https://github.com/prrvchr/OAuth2OOo.git"
  branch:   "master"
  commit:   "9a9931f"
git-subrepo:
  version:  "0.4.0"
  origin:   "https://github.com/ingydotnet/git-subrepo"
  commit:   "5d6aba9"

+++ OAuth2OOo/python/database/dbtools.py
@@ -318,13 +318,14 @@ def getRowResult(result, index=(0,), separator=' '):
     return tuple(sequence)
 
 def getValueFromResult(result, index, default=None):
+    # TODO: 'TINYINT' is buggy: don't use it
     dbtype = result.MetaData.getColumnTypeName(index)
     if dbtype == 'VARCHAR':
         value = result.getString(index)
     elif dbtype == 'BOOLEAN':
         value = result.getBoolean(index)
     elif dbtype == 'TINYINT':
-        value = result.getByte(index)
+        value = result.getShort(index)
     elif dbtype == 'SMALLINT':
         value = result.getShort(index)
     elif dbtype == 'INTEGER':
